Extract labels filter dropdown to own component
